+++ vpc-20160428/src/Models/DescribeGrantRulesToEcrResponseBody.php
@@ -1,0 +1,136 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Vpc\V20160428\Models;
+
+use AlibabaCloud\Dara\Model;
+use AlibabaCloud\SDK\Vpc\V20160428\Models\DescribeGrantRulesToEcrResponseBody\ecrGrantRules;
+
+class DescribeGrantRulesToEcrResponseBody extends Model
+{
+    /**
+     * @var int
+     */
+    public $count;
+
+    /**
+     * @var ecrGrantRules[]
+     */
+    public $ecrGrantRules;
+
+    /**
+     * @var int
+     */
+    public $pageNumber;
+
+    /**
+     * @var int
+     */
+    public $pageSize;
+
+    /**
+     * @var string
+     */
+    public $requestId;
+
+    /**
+     * @var int
+     */
+    public $totalCount;
+    protected $_name = [
+        'count' => 'Count',
+        'ecrGrantRules' => 'EcrGrantRules',
+        'pageNumber' => 'PageNumber',
+        'pageSize' => 'PageSize',
+        'requestId' => 'RequestId',
+        'totalCount' => 'TotalCount',
+    ];
+
+    public function validate()
+    {
+        if (\is_array($this->ecrGrantRules)) {
+            Model::validateArray($this->ecrGrantRules);
+        }
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->count) {
+            $res['Count'] = $this->count;
+        }
+
+        if (null !== $this->ecrGrantRules) {
+            if (\is_array($this->ecrGrantRules)) {
+                $res['EcrGrantRules'] = [];
+                $n1 = 0;
+                foreach ($this->ecrGrantRules as $item1) {
+                    $res['EcrGrantRules'][$n1] = null !== $item1 ? $item1->toArray($noStream) : $item1;
+                    ++$n1;
+                }
+            }
+        }
+
+        if (null !== $this->pageNumber) {
+            $res['PageNumber'] = $this->pageNumber;
+        }
+
+        if (null !== $this->pageSize) {
+            $res['PageSize'] = $this->pageSize;
+        }
+
+        if (null !== $this->requestId) {
+            $res['RequestId'] = $this->requestId;
+        }
+
+        if (null !== $this->totalCount) {
+            $res['TotalCount'] = $this->totalCount;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['Count'])) {
+            $model->count = $map['Count'];
+        }
+
+        if (isset($map['EcrGrantRules'])) {
+            if (!empty($map['EcrGrantRules'])) {
+                $model->ecrGrantRules = [];
+                $n1 = 0;
+                foreach ($map['EcrGrantRules'] as $item1) {
+                    $model->ecrGrantRules[$n1] = ecrGrantRules::fromMap($item1);
+                    ++$n1;
+                }
+            }
+        }
+
+        if (isset($map['PageNumber'])) {
+            $model->pageNumber = $map['PageNumber'];
+        }
+
+        if (isset($map['PageSize'])) {
+            $model->pageSize = $map['PageSize'];
+        }
+
+        if (isset($map['RequestId'])) {
+            $model->requestId = $map['RequestId'];
+        }
+
+        if (isset($map['TotalCount'])) {
+            $model->totalCount = $map['TotalCount'];
+        }
+
+        return $model;
+    }
+}